Avoid writing padding bytes if no content is inflated
Port the change for Chromium issue 1302606 which avoids writing padding bytes for debug purposes except for debug builds. In particular, this ensures that the JDK `InflaterInputStream::read(byte[] b, int off, int len)` contract can be satisfied by this implementation. See:

- https://bugs.chromium.org/p/chromium/issues/detail?id=1302606
- https://chromium.googlesource.com/chromium/src/+/36be3c7a465f00f4edefccc10bba00bbebe1843d%5E%21/

This is complementary to the work Volker Simonis is doing upstream on making `InflaterInputStream` more tolerant of this kind of behavior, providing compatibility for earlier JDK releases.

- https://bugs.openjdk.org/browse/JDK-8281962
- https://bugs.openjdk.org/browse/JDK-8283758

Signed-off-by: Danny Thomas <dannyt@netflix.com>